% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pleather.R
\name{pleather}
\alias{pleather}
\title{pleather}
\usage{
pleather(
  dat,
  treat.id,
  iterations,
  subsample,
  pred.vars,
  preds.op,
  dep.var,
  unit.var,
  time.var,
  pre.time,
  opt.time,
  unit.var.name,
  plot.time,
  synth.opt.method,
  plot.title,
  plot.y,
  plot.x,
  plot.y.int,
  plot.x.int,
  treated.name,
  treated.color,
  save.to
)
}
\arguments{
\item{dat}{A data.frame object ready to pass into Synth}

\item{treat.id}{The ID for the treated unit - this comes from the unit.var column of dat}

\item{iterations}{The number of times that pleather resamples from the donor pool and re-runs Synth}

\item{subsample}{The proportion of the donor pool to be randomly chosen for each iteration}

\item{pred.vars}{The names of the predictor variables for Synth}

\item{preds.op}{The operator to be used on the predictors - the default is 'mean'}

\item{dep.var}{The name of the dependent variable}

\item{unit.var}{The name of the column of dat that contains the ID for Synth}

\item{time.var}{The name of the column of dat that contains the time variable}

\item{pre.time}{The period of time before the treatment takes effect}

\item{opt.time}{The period of time over which Synth will optimize the pre-event fit}

\item{unit.var.name}{The name of the column of dat that contains the names of the units}

\item{plot.time}{The period of time to be plotted by Synth}

\item{synth.opt.method}{The optimization method for Synth to use - the default is All}

\item{plot.title}{The title for the ggplot created by pleather}

\item{plot.y}{The label for the y-axis of the ggplot}

\item{plot.x}{The label for the x-axis of the ggplot}

\item{plot.y.int}{The y-intercept for a horizontal line - do not include if unwanted}

\item{plot.x.int}{The x-intercept for a vertical line - do not include if unwanted}

\item{treated.name}{The name of the treated unit as you want it to appear in the ggplot legend}

\item{treated.color}{The color of the line for the treated unit in the ggplot - default is red4 - Go Hokies!}

\item{save.to}{The filepath for where you want the figure saved if desired}
}
\value{
A data.frame object which contains data on the means and standard deviations from the bootstrapped synthetic control method
}
\description{
Performs the bootstrapped synthetic controls method using the Synth package
}
\examples{
data(fake)
boot.df <- pleather(dat = fake, treat.id = 21, iterations = 50, subsample = .5, pred.vars = c('Effort', 'Funding'), preds.op = 'mean', dep.var = 'Output', unit.var = 'ID', time.var = 'Month', pre.time = c(1:24), opt.time = c(1:24), unit.var.name = 'Name', plot.time = c(1:36), synth.opt.method = 'BFGS', plot.title = 'Example Using Bootstapped SCM', plot.y = 'Output', plot.x = 'Month', plot.x.int = 24, treated.name = 'Michael', treated.color = 'red4', save.to = 'pleather_example_figure.png')

}
